+++ .gitignore
@@ -60,5 +60,5 @@ typings/
 .env
 
 .idea
-
+package-lock.json
