+++ extension/manifest.json
@@ -2,7 +2,7 @@
   "manifest_version": 2,
   "name": "__MSG_extensionName__",
   "short_name": "__MSG_extensionName__",
-  "version": "0.1.4",
+  "version": "0.1.5",
   "description": "__MSG_extensionDescription__",
   "default_locale": "en",
   "icons": {
@@ -44,7 +44,7 @@
   "content_scripts": [
     {
       "matches": [ "<all_urls>" ],
-      "js": ["scripts/whatsElement.js","scripts/content_script.js"],
+      "js": ["scripts/whatsElement.min.js","scripts/content_script.js"],
       "all_frames": true
     }
   ]
