+++ extension/scripts/helper.js
@@ -1,0 +1,1 @@
+"use strict";function createElement(e,t){var o=document.createElement(e||"div");return t&&(o.id=t),o}function getCoords(e){var t=e.getBoundingClientRect(),o=document.body,n=document.documentElement,l=window.pageYOffset||n.scrollTop||o.scrollTop,r=window.pageXOffset||n.scrollLeft||o.scrollLeft,d=n.clientTop||o.clientTop||0,c=n.clientLeft||o.clientLeft||0,u=t.top+l-d,f=t.left+r-c;return{top:Math.round(u),left:Math.round(f)}}Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _default={createElement:createElement,getCoords:getCoords};exports.default=_default;
