--- extension/scripts/whatsElement.js
@@ -1,1 +1,0 @@
-!function(e,t){function n(t){var n=t.getBoundingClientRect(),o=r.body,i=r.documentElement,l=e.pageYOffset||i.scrollTop||o.scrollTop,d=e.pageXOffset||i.scrollLeft||o.scrollLeft,a=i.clientTop||o.clientTop||0,p=i.clientLeft||o.clientLeft||0,c=n.top+l-a,u=n.left+d-p;return{top:Math.round(c),left:Math.round(u)}}function o(){r.querySelectorAll("."+l+"-active").forEach(function(e){e.classList.remove(l+"-active")})}function i(e,t){var n=r.createElement(e||"div");return t&&(n.id=t),n}var r=e.document,l="whats-element",d=l+"-tip-container",a=l+"-tip-delete",p=l+"-copy",c=l+"-unique-container",u=l+"-unique-id",s=function(e){this.options=Object.assign({},{draw:!0,simpleId:!0},e),this.lastClick=r.body;var t=this;r.addEventListener("mousedown",function(e){t.lastClick=e.target,t.focusedElement!==t.lastClick&&f.clean()})},f=s.prototype;f.getUniqueId=function(e,t){if(!((e=e||this.lastClick)instanceof HTMLElement))return console.error("invalid input,not a HTML element"),null;var o={wid:"",type:"",top:n(e).top,left:n(e).left,viewLeft:e.getBoundingClientRect().left,viewTop:e.getBoundingClientRect().top,text:e.innerText},i=e.id,l=e.name,d=e.tagName.toLowerCase(),a=e.type?e.type.toLowerCase():"",p="",c=e.classList||[];if(c.forEach(function(e){p+="."+e}),"body"!==d&&"html"!==d||(o.wid=d,o.type=d),i&&r.getElementById(i)===e){var u=new RegExp("^[a-zA-Z]+");!t&&this.options.simpleId?o.wid=i:u.test(i)&&(o.wid=d+"#"+i),o.type="document.getElementById()"}if(!o.wid&&l&&r.getElementsByName(l)[0]===e&&(o.wid=l,o.type="document.getElementsByName()"),!o.wid&&p&&r.querySelector(d+p)===e){o.wid=d+p,o.type="document.querySelector()";var y=c.length;if(y>2){for(var g=1,h=[];Math.pow(2,g)<y;)h.push(Math.pow(2,g)),g++;h.push(y);for(var m=0;m<h.length;m++)h[m]}}if("radio"===a){var w=d+"[value='"+e.value+"']";l&&(w+="[name='"+l+"']"),r.querySelector(w)===e&&(o.wid=w,o.type="document.querySelector()")}if(!o.wid){w=d;w=p?w+p:w,w=l?w+"[name='"+l+"']":w,f.getTarget(w)===e&&(o.wid=w,o.type="document.querySelector()")}if(!o.wid){w=d;w=p?w+p:w;var v=r.querySelectorAll(w);if(v&&v.length>0){for(var x=null,b=0;b<v.length;b++)if(e===v[b]){x=b+1;break}x&&(w=w+":nth-child("+x+")",r.querySelector(w)===e&&(o.wid=w,o.type="document.querySelector()"))}}if(!o.wid){if(!e.parentNode)return;var C=s.prototype.getUniqueId(e.parentNode,!0),E=C?C.wid:"";if(!E)return{wid:"",type:"NO_LOCATION"};var q=d;p&&(q+=p);w=E+">"+q;if(r.querySelectorAll(w).length>1){w=null;x=null;for(var T=0;T<e.parentNode.children.length;T++)if(e.parentNode.children[T]===e){x=T+1;break}if(x>=1){w=E+">"+q+":nth-child("+x+")";r.querySelector(w)!=e&&(w=null)}}o.wid=w,o.type="document.querySelector()"}return this.focusedElement=f.getTarget(o.wid),!t&&this.options.draw&&s.prototype.draw(o),o},f.getTarget=function(e){return r.getElementById(e)||r.getElementsByName(e)[0]||r.querySelector(e)},f.draw=function(t){var n=f.getTarget(t.wid);if(n){var l=r.getElementById(d)?r.getElementById(d):i("aside",d);l.innerHTML="",l.addEventListener("mousedown",function(e){e.stopPropagation()});var y=i("div",a);y.innerText="x",y.onclick=function(e){e.stopPropagation(),s.prototype.clean()},l.appendChild(y);var g=i("div",c),h=i("input",u);h.readOnly=!0,h.className=t.type,h.value=t.wid;var m=i("div",p);m.setAttribute("query-type",t.type),m.innerText="复制",m.onclick=function(){h.select(),r.execCommand("Copy")},g.appendChild(h),g.appendChild(m),l.appendChild(g);var w=n.getBoundingClientRect().top+n.offsetHeight,v=n.getBoundingClientRect().left,x=v<100?v:v+n.offsetWidth/2-(240|l.offsetWidth)/2-r.body.getBoundingClientRect().left;l.style.left=x+"px",l.style.top=w+e.scrollY+10+"px",r.body.appendChild(l),o(),n.classList.add("whats-element-active")}else console.error("no this element:"+t.wid)},f.clean=function(){o();var e=r.getElementById(d);e&&(e.outerHTML="")},void 0!==e&&null!==e&&(e.whatsElement=s),void 0!==e&&null!==e||(this.whatsElement=s),"undefined"!=typeof module&&module.exports&&(module.exports=s),"function"==typeof define&&define(function(){return s});var y=i("style"),g="#"+d+"{position:absolute;color:#8ed3fb;font-size:14px;z-index:1000;background-color:rgba(255, 255, 255,0.9);box-sizing:border-box;box-shadow:rgba(0, 0, 0, 0.2) 0px 1px 10px 3px;padding: 10px 20px;border-radius: 36px;}";g+="#"+a+"{cursor:pointer;position:absolute;top:-10px;width:20px;height: 20px;left:calc(50% - 8px);text-align:center;clip-path: polygon(50% 0,100% 50%,50% 100%,0 50%);background:#fff;}",g+="#"+a+":hover{background:#fffbf0}",g+="#"+c+"{display:flex;justify-content:space-around;}",g+="#"+u+"{border:1px solid #d1d5da;box-shadow:inset 0 1px 2px rgba(27,31,35,0.075);text-indent:10px;}",g+="#"+p+"{background:aliceblue;cursor: pointer;}",g+="#"+p+"::after{position:absolute;z-index:1000000;padding: 0.1em 0.75em;color: #fff;text-align:center;text-shadow:none;text-transform:none;content:attr(query-type);background: #a88462;border-radius: 3px;opacity: 0;transition:all .5s}",g+="#"+p+":hover:after{display:inline-block;opacity: 1;}",g+=".whats-element-active{outline:red 1px dashed !important}",y.innerText=g,r.head.appendChild(y)}(window);
