+++ gulpfile.js
@@ -1,25 +1,48 @@
 var gulp = require('gulp');
+var babelify = require('babelify')
+var source = require('vinyl-source-stream');
+var buffer = require('vinyl-buffer');
 var uglify = require('gulp-uglify');
 var pump = require('pump');
-var watch = require('gulp-watch');
 var gulpZip = require('gulp-zip');
 var rename  = require('gulp-rename');
 var fs = require('fs');
+var browserify = require('browserify');
 const packageJson = JSON.parse(fs.readFileSync('./package.json'));
 const version = packageJson.version;
 gulp.task('compress', function (cb) {
-    pump([
-            gulp.src('src/**/*.js'),
-            uglify(),
-            gulp.dest('dist'),
-            gulp.dest('extension/scripts'),
-            rename({
-              suffix:".min"
-            }),
-            gulp.dest('dist'),
-        ],
-        cb
-    );
+    var b = browserify({
+        entries:"src/whatsElement.js"
+    })
+
+    return b
+            .transform(babelify)
+            .bundle()
+            .pipe(source("whatsElement.js"))
+            .pipe(buffer())
+            .pipe(gulp.dest('dist'))
+            .pipe(uglify())
+            .pipe(rename({
+                suffix:".min"
+            }))
+            .pipe(gulp.dest('dist'))
+            .pipe(gulp.dest('extension/scripts'))
+
+
+
+    // pump([
+    //         gulp.src('src/**/whatsElement.js'),
+    //         babel(),
+    //         uglify(),
+    //         gulp.dest('dist'),
+    //         gulp.dest('extension/scripts'),
+    //         rename({
+    //           suffix:".min"
+    //         }),
+    //         gulp.dest('dist'),
+    //     ],
+    //     cb
+    // );
 });
 
 gulp.task('watch', function () {
